feat: makes Vinto respond to users leaving
Because we don't want leftover rectangles displayed for participants who
have left. The app should re-render and only display participants in the
conference.

HOW it WORKS: respond to USER_LEFT event with respondToUserLeftHandler
which uses setTracks and is defined on the application component and is
passed to jitsiConnector through the connect function.